Per-window SourceViewPanel and Tab visitors
In order to fix an issue with closing secondary windows with changed,
inactive 'SourceViewPanel's we need a way to visit the tabs of just a
specific Window. The current 'SecondaryAllowClose' logic that relies
on checking 'mWidgetWindow' is broken. This commit just introduces the
needed utility methods and refactors the all-window methods to use them.