Allow secondary windows to be maximized
Prior to this change, secondary windows could be minimized by
right-clicking on the title bar, but not maximized. Maximizing is
very useful for multi-monitor setups. The popup menu includes restore
as well. To enable these actions directly from the titlebar without
navigating a popup, the 'ToolWindow' style would have to be removed
because it inhibits the standard system buttons.